Added Various Features
New Features:
- add employee (post)
- get employee by Id (get)
- update employee (put)
- delete employee (delete)

+++ src/main/java/com/csa/samplefullstack/entity/Employee.java
@@ -1,6 +1,11 @@
 package com.csa.samplefullstack.entity;
 
-import javax.persistence.*;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import javax.persistence.GenerationType;
 
 @Entity
 @Table(name="tbl_employee")
@@ -8,12 +13,12 @@ public class Employee {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int employeeId;
+	private Long employeeId;
 	private String name;
 	private String department;
 	private String location;
 
-	public Employee (int employeeId, String name, String department, String location){
+	public Employee (Long employeeId, String name, String department, String location){
 		this.employeeId = employeeId;
 		this.name = name;
 		this.department = department;
@@ -22,7 +27,7 @@ public class Employee {
 	public Employee(){
 
 	}
-	public void setEmployeeId(int employeeId){
+	public void setEmployeeId(Long employeeId){
 		this.employeeId = employeeId;
 	}
 	public void setName(String name){
@@ -34,7 +39,7 @@ public class Employee {
 	public void setLocation(String location){
 		this.location = location;
 	}
-	public int getEmployeeId() {
+	public Long getEmployeeId() {
 		return this.employeeId;
 	}
 	public String getName() {
